requirements.txt
asgiref==3.5.2
black==22.3.0
certifi==2022.5.18.1
click==8.1.3
colorama==0.4.4
Django==4.0.5
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytz==2022.1
sentry-sdk==1.5.12
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
